chore: cleanup

+++ src/JKA_YBEProxy/EnginePatch/Proxy_sv_client.cpp
@@ -462,7 +462,7 @@ void Proxy_SV_UserinfoChanged(client_t* cl) {
 	{
 		const size_t modelLen = strlen(val);
 
-		if (!Q_IsValidAsciiStr(val) || modelLen >= /*MAX_QPATH*/ proxy.cvars.proxy_sv_modelPathLength.integer)
+		if (!Q_IsValidAsciiStr(val) || modelLen >= /*MAX_QPATH*/ (size_t)proxy.cvars.proxy_sv_modelPathLength.integer)
 		{
 			Info_SetValueForKey(cl->userinfo, "model", "kyle");
 		}
